# Code/Servidor/swagger_server/models/song_item.py
# coding: utf-8

from __future__ import absolute_import
from datetime import date, datetime  # noqa: F401

from typing import List, Dict  # noqa: F401

from swagger_server.models.base_model_ import Model
from swagger_server import util


class SongItem(Model):
    """NOTE: This class is auto generated by the swagger code generator program.

    Do not edit the class manually.
    """

    def __init__(self, id: str=None, name: str=None, lenght: str=None, author_id: str=None, author_name: str=None, album_id: str=None, album_name: str=None, genre: List[str]=None):  # noqa: E501
        """SongItem - a model defined in Swagger

        :param id: The id of this SongItem.  # noqa: E501
        :type id: str
        :param name: The name of this SongItem.  # noqa: E501
        :type name: str
        :param lenght: The lenght of this SongItem.  # noqa: E501
        :type lenght: str
        :param author_id: The author_id of this SongItem.  # noqa: E501
        :type author_id: str
        :param author_name: The author_name of this SongItem.  # noqa: E501
        :type author_name: str
        :param album_id: The album_id of this SongItem.  # noqa: E501
        :type album_id: str
        :param album_name: The album_name of this SongItem.  # noqa: E501
        :type album_name: str
        :param genre: The genre of this SongItem.  # noqa: E501
        :type genre: List[str]
        """
        self.swagger_types = {
            'id': str,
            'name': str,
            'lenght': str,
            'author_id': str,
            'author_name': str,
            'album_id': str,
            'album_name': str,
            'genre': List[str]
        }

        self.attribute_map = {
            'id': 'id',
            'name': 'name',
            'lenght': 'lenght',
            'author_id': 'authorID',
            'author_name': 'authorName',
            'album_id': 'albumID',
            'album_name': 'albumName',
            'genre': 'genre'
        }

        self._id = id
        self._name = name
        self._lenght = lenght
        self._author_id = author_id
        self._author_name = author_name
        self._album_id = album_id
        self._album_name = album_name
        self._genre = genre

    @classmethod
    def from_dict(cls, dikt) -> 'SongItem':
        """Returns the dict as a model

        :param dikt: A dict.
        :type: dict
        :return: The SongItem of this SongItem.  # noqa: E501
        :rtype: SongItem
        """
        return util.deserialize_model(dikt, cls)

    @property
    def id(self) -> str:
        """Gets the id of this SongItem.


        :return: The id of this SongItem.
        :rtype: str
        """
        return self._id

    @id.setter
    def id(self, id: str):
        """Sets the id of this SongItem.


        :param id: The id of this SongItem.
        :type id: str
        """
        if id is None:
            raise ValueError("Invalid value for `id`, must not be `None`")  # noqa: E501

        self._id = id

    @property
    def name(self) -> str:
        """Gets the name of this SongItem.


        :return: The name of this SongItem.
        :rtype: str
        """
        return self._name

    @name.setter
    def name(self, name: str):
        """Sets the name of this SongItem.


        :param name: The name of this SongItem.
        :type name: str
        """
        if name is None:
            raise ValueError("Invalid value for `name`, must not be `None`")  # noqa: E501

        self._name = name

    @property
    def lenght(self) -> str:
        """Gets the lenght of this SongItem.


        :return: The lenght of this SongItem.
        :rtype: str
        """
        return self._lenght

    @lenght.setter
    def lenght(self, lenght: str):
        """Sets the lenght of this SongItem.


        :param lenght: The lenght of this SongItem.
        :type lenght: str
        """
        if lenght is None:
            raise ValueError("Invalid value for `lenght`, must not be `None`")  # noqa: E501

        self._lenght = lenght

    @property
    def author_id(self) -> str:
        """Gets the author_id of this SongItem.


        :return: The author_id of this SongItem.
        :rtype: str
        """
        return self._author_id

    @author_id.setter
    def author_id(self, author_id: str):
        """Sets the author_id of this SongItem.


        :param author_id: The author_id of this SongItem.
        :type author_id: str
        """
        if author_id is None:
            raise ValueError("Invalid value for `author_id`, must not be `None`")  # noqa: E501

        self._author_id = author_id

    @property
    def author_name(self) -> str:
        """Gets the author_name of this SongItem.


        :return: The author_name of this SongItem.
        :rtype: str
        """
        return self._author_name

    @author_name.setter
    def author_name(self, author_name: str):
        """Sets the author_name of this SongItem.


        :param author_name: The author_name of this SongItem.
        :type author_name: str
        """
        if author_name is None:
            raise ValueError("Invalid value for `author_name`, must not be `None`")  # noqa: E501

        self._author_name = author_name

    @property
    def album_id(self) -> str:
        """Gets the album_id of this SongItem.


        :return: The album_id of this SongItem.
        :rtype: str
        """
        return self._album_id

    @album_id.setter
    def album_id(self, album_id: str):
        """Sets the album_id of this SongItem.


        :param album_id: The album_id of this SongItem.
        :type album_id: str
        """
        if album_id is None:
            raise ValueError("Invalid value for `album_id`, must not be `None`")  # noqa: E501

        self._album_id = album_id

    @property
    def album_name(self) -> str:
        """Gets the album_name of this SongItem.


        :return: The album_name of this SongItem.
        :rtype: str
        """
        return self._album_name

    @album_name.setter
    def album_name(self, album_name: str):
        """Sets the album_name of this SongItem.


        :param album_name: The album_name of this SongItem.
        :type album_name: str
        """
        if album_name is None:
            raise ValueError("Invalid value for `album_name`, must not be `None`")  # noqa: E501

        self._album_name = album_name

    @property
    def genre(self) -> List[str]:
        """Gets the genre of this SongItem.


        :return: The genre of this SongItem.
        :rtype: List[str]
        """
        return self._genre

    @genre.setter
    def genre(self, genre: List[str]):
        """Sets the genre of this SongItem.


        :param genre: The genre of this SongItem.
        :type genre: List[str]
        """
        if genre is None:
            raise ValueError("Invalid value for `genre`, must not be `None`")  # noqa: E501

        self._genre = genre