# Code/Servidor/swagger_server/swagger/swagger.yaml
...
swagger: "2.0"
info:
  description: "Back-end API del Proyecto Cierzo"
  version: "2.0.0"
  title: "Proyecto Cierzo API"
  license:
    name: "Apache 2.0"
    url: "http://www.apache.org/licenses/LICENSE-2.0.html"
host: "localhost:8080"
basePath: "/api/"
...
schemes:
- "https"
paths:
  /songs:
    get:
      tags:
      - "public"
      summary: "busca canciones con ciertos parámetros"
      description: "Al pasarle ciertos parámetros devuelve cancionese que se ajusten\
        \ a ellos\n"
      operationId: "search_song"
      produces:
      - "application/json"
      parameters:
      - name: "name"
        in: "query"
        description: "nombre de la canción"
        required: false
        type: "string"
      - name: "author"
        in: "query"
        description: "autor de la canción"
        required: false
        type: "string"
      - name: "genre"
        in: "query"
        description: "genero de la canción"
        required: false
        type: "string"
      - name: "skip"
        in: "query"
        description: "number of records to skip for pagination"
        required: false
        type: "integer"
        minimum: 0
        format: "int32"
      - name: "limit"
        in: "query"
        description: "maximum number of records to return"
        required: false
        type: "integer"
        maximum: 50.0
        minimum: 0
        format: "int32"
      responses:
        200:
          description: "search results matching criteria"
          schema:
            type: "array"
            items:
              $ref: "#/definitions/SongItem"
        400:
          description: "bad input parameter"
      x-swagger-router-controller: "swagger_server.controllers.public_controller"
    post:
      tags:
      - "admins"
      summary: "añade una canción"
      description: "Un admin añade una canción a la base de datos."
      operationId: "add_song"
      consumes:
      - "application/json"
      produces:
      - "application/json"
      parameters:
      - in: "body"
        name: "songItem"
        description: "Song item to add"
        required: false
        schema:
          $ref: "#/definitions/SongItemNew"
      responses:
        201:
          description: "item created"
          schema:
            $ref: "#/definitions/SongItem"
        400:
          description: "invalid input, object invalid"
        401:
          description: "need authentification"
        403:
          description: "operation forbidden for user"
      x-swagger-router-controller: "swagger_server.controllers.admins_controller"
  /songs/{songID}:
    get:
      tags:
      - "public"
      summary: "obtiene información de una canción"
      description: "Obtiene los datos de la canción identificada por songID\n"
      operationId: "get_song"
      produces:
      - "application/json"
      parameters:
      - name: "songID"
        in: "path"
        description: "ID de la canción"
        required: true
        type: "string"
      responses:
        200:
          description: "success"
          schema:
            $ref: "#/definitions/SongItem"
        400:
          description: "bad input parameter"
        404:
          description: "not found"
      x-swagger-router-controller: "swagger_server.controllers.public_controller"
    put:
      tags:
      - "admins"
      summary: "actualiza la información de una canción"
      description: "Un admin actualiza la información de una canción."
      operationId: "update_song"
      consumes:
      - "application/json"
      produces:
      - "application/json"
      parameters:
      - name: "songID"
        in: "path"
        description: "ID de la canción"
        required: true
        type: "string"
      - in: "body"
        name: "songItem"
        description: "Song item to update"
        required: false
        schema:
          $ref: "#/definitions/SongItemNew"
      responses:
        200:
          description: "item updated"
          schema:
            $ref: "#/definitions/SongItem"
        400:
          description: "invalid input, object invalid"
        401:
          description: "need authentification"
        403:
          description: "operation forbidden for user"
        404:
          description: "not found"
      x-swagger-router-controller: "swagger_server.controllers.admins_controller"
    delete:
      tags:
      - "admins"
      summary: "elimina una canción"
      description: "Elimina información y fichero de  la canción identificada por\
        \ songID\n"
      operationId: "delete_song"
      produces:
      - "application/json"
      parameters:
      - name: "songID"
        in: "path"
        description: "ID de la canción"
        required: true
        type: "string"
      responses:
        200:
          description: "song deleted"
        400:
          description: "bad input parameter"
        401:
          description: "need authentification"
        403:
          description: "operation forbidden for user"
        404:
          description: "not found"
      x-swagger-router-controller: "swagger_server.controllers.admins_controller"
  /songs/{songID}/file:
    get:
      tags:
      - "public"
      summary: "obtiene el archivo de audio de una canción"
      description: "Obtiene el archivo de audio de la canción identificada por songID\n"
      operationId: "get_song_file"
      produces:
      - "application/vnd.mp3"
      parameters:
      - name: "songID"
        in: "path"
        description: "ID de la canción"
        required: true
        type: "string"
      responses:
        200:
          description: "success"
          schema:
            type: "file"
        400:
          description: "bad input parameter"
        404:
          description: "not found"
      x-swagger-router-controller: "swagger_server.controllers.public_controller"
    post:
      tags:
      - "admins"
      summary: "carga el archivo de audio de una canción"
      description: "Un admin carga el archivo de audio de una canción con identificador\
        \ songID."
      operationId: "upload_song_file"
      consumes:
      - "multipart/form-data"
      produces:
      - "application/json"
      parameters:
      - name: "songID"
        in: "path"
        description: "ID de la canción"
        required: true
        type: "string"
      - name: "songFile"
        in: "formData"
        description: "Song item to update"
        required: true
        type: "file"
      responses:
        201:
          description: "file uploaded"
        400:
          description: "invalid input, object invalid"
        401:
          description: "need authentification"
        403:
          description: "operation forbidden for user"
        404:
          description: "not found"
      x-swagger-router-controller: "swagger_server.controllers.admins_controller"
    delete:
      tags:
      - "admins"
      summary: "elimina el archivo de audio de una canción"
      description: "Elimina el fichero de audio de una canción identificada por songID\n"
      operationId: "delete_song_file"
      produces:
      - "application/json"
      parameters:
      - name: "songID"
        in: "path"
        description: "ID de la canción"
        required: true
        type: "string"
      responses:
        200:
          description: "song deleted"
        400:
          description: "bad input parameter"
        401:
          description: "need authentification"
        403:
          description: "operation forbidden for user"
        404:
          description: "not found"
      x-swagger-router-controller: "swagger_server.controllers.admins_controller"
  /songs/{songID}/image:
    get:
      tags:
      - "public"
      summary: "obtiene la carátula de una canción"
      description: "Obtiene la carátula de la canción identificada por songID\n"
      operationId: "get_song_image"
      produces:
      - "image/png"
      parameters:
      - name: "songID"
        in: "path"
        description: "ID de la canción"
        required: true
        type: "string"
      responses:
        200:
          description: "success"
          schema:
            type: "file"
        400:
          description: "bad input parameter"
        404:
          description: "not found"
      x-swagger-router-controller: "swagger_server.controllers.public_controller"
    post:
      tags:
      - "admins"
      summary: "carga la carátula de una canción"
      description: "Un admin carga la carátula de una canción con identificador songID."
      operationId: "upload_song_image"
      consumes:
      - "multipart/form-data"
      produces:
      - "application/json"
      parameters:
      - name: "songID"
        in: "path"
        description: "ID de la canción"
        required: true
        type: "string"
      - name: "songImage"
        in: "formData"
        description: "Image item to update"
        required: true
        type: "file"
      responses:
        201:
          description: "file uploaded"
        400:
          description: "invalid input, object invalid"
        401:
          description: "need authentification"
        403:
          description: "operation forbidden for user"
        404:
          description: "not found"
      x-swagger-router-controller: "swagger_server.controllers.admins_controller"
    delete:
      tags:
      - "admins"
      summary: "elimina la carátula de una canción"
      description: "Elimina la carátula de una canción identificada por songID\n"
      operationId: "delete_song_image"
      produces:
      - "application/json"
      parameters:
      - name: "songID"
        in: "path"
        description: "ID de la canción"
        required: true
        type: "string"
      responses:
        200:
          description: "image deleted"
        400:
          description: "bad input parameter"
        401:
          description: "need authentification"
        403:
          description: "operation forbidden for user"
        404:
          description: "not found"
      x-swagger-router-controller: "swagger_server.controllers.admins_controller"
  /albums:
    get:
      tags:
      - "public"
      summary: "busca álbunes con ciertos parámetros"
      description: "Al pasarle ciertos parámetros devuelve álbunes que se ajusten\
        \ a ellos\n"
      operationId: "search_album"
      produces:
      - "application/json"
      parameters:
      - name: "name"
        in: "query"
        description: "nombre del álbum"
        required: false
        type: "string"
      - name: "author"
        in: "query"
        description: "autor del álbum"
        required: false
        type: "string"
      - name: "skip"
        in: "query"
        description: "number of records to skip for pagination"
        required: false
        type: "integer"
        minimum: 0
        format: "int32"
      - name: "limit"
        in: "query"
        description: "maximum number of records to return"
        required: false
        type: "integer"
        maximum: 50.0
        minimum: 0
        format: "int32"
      responses:
        200:
          description: "search results matching criteria"
          schema:
            type: "array"
            items:
              $ref: "#/definitions/AlbumItem"
        400:
          description: "bad input parameter"
      x-swagger-router-controller: "swagger_server.controllers.public_controller"
    post:
      tags:
      - "admins"
      summary: "crea un álbum"
      description: "Un administrador crea un álbum"
      operationId: "add_album"
      consumes:
      - "application/json"
      produces:
      - "application/json"
      parameters:
      - in: "body"
        name: "albumItem"
        description: "Album item to add"
        required: false
        schema:
          $ref: "#/definitions/AlbumItemNew"
      responses:
        201:
          description: "item created"
          schema:
            $ref: "#/definitions/AlbumItem"
        400:
          description: "invalid input, object invalid"
        401:
          description: "need authentification"
        403:
          description: "operation forbiden for user"
      x-swagger-router-controller: "swagger_server.controllers.admins_controller"
  /albums/{albumID}:
    get:
      tags:
      - "public"
      summary: "obtiene un álbum"
      description: "Obtiene los datos del álbum identificada por albumID\n"
      operationId: "get_album"
      produces:
      - "application/json"
      parameters:
      - name: "albumID"
        in: "path"
        description: "ID del álbum"
        required: true
        type: "string"
      responses:
        200:
          description: "success"
          schema:
            $ref: "#/definitions/AlbumItem"
        400:
          description: "bad input parameter"
        404:
          description: "not found"
      x-swagger-router-controller: "swagger_server.controllers.public_controller"
    delete:
      tags:
      - "admins"
      summary: "elimina un álbum"
      description: "Elimina el álbum identificado por albumID\n"
      operationId: "delete_album"
      produces:
      - "application/json"
      parameters:
      - name: "albumID"
        in: "path"
        description: "ID del álbum"
        required: true
        type: "string"
      responses:
        200:
          description: "album deleted"
        400:
          description: "bad input parameter"
        401:
          description: "need authentification"
        403:
          description: "operation forbidden for user"
        404:
          description: "not found"
      x-swagger-router-controller: "swagger_server.controllers.admins_controller"
  /albums/{albumID}/image:
    get:
      tags:
      - "public"
      summary: "obtiene la carátula de un álbum"
      description: "Obtiene la carátula de un álbum identificado por albumID\n"
      operationId: "get_album_image"
      produces:
      - "image/png"
      parameters:
      - name: "albumID"
        in: "path"
        description: "ID del álbum"
        required: true
        type: "string"
      responses:
        200:
          description: "success"
          schema:
            type: "file"
        400:
          description: "bad input parameter"
        404:
          description: "not found"
      x-swagger-router-controller: "swagger_server.controllers.public_controller"
    post:
      tags:
      - "admins"
      summary: "carga la carátula de un álbum"
      description: "Un admin carga la carátula de un álbum con identificador albumID."
      operationId: "upload_album_image"
      consumes:
      - "multipart/form-data"
      produces:
      - "application/json"
      parameters:
      - name: "albumID"
        in: "path"
        description: "ID del álbum"
        required: true
        type: "string"
      - name: "albumImage"
        in: "formData"
        description: "Album image to update"
        required: true
        type: "file"
      responses:
        201:
          description: "image uploaded"
        400:
          description: "invalid input, object invalid"
        401:
          description: "need authentification"
        403:
          description: "operation forbidden for user"
        404:
          description: "not found"
      x-swagger-router-controller: "swagger_server.controllers.admins_controller"
    delete:
      tags:
      - "admins"
      summary: "elimina la carátula de un álbum"
      description: "Elimina la carátula de un álbum identificadoa por albumID\n"
      operationId: "delete_album_image"
      produces:
      - "application/json"
      parameters:
      - name: "albumID"
        in: "path"
        description: "ID del álbum"
        required: true
        type: "string"
      responses:
        200:
          description: "image deleted"
        400:
          description: "bad input parameter"
        401:
          description: "need authentification"
        403:
          description: "operation forbidden for user"
        404:
          description: "not found"
      x-swagger-router-controller: "swagger_server.controllers.admins_controller"
  /albums/{albumID}/songs:
    post:
      tags:
      - "admins"
      summary: "añade una canción a un album"
      description: "Un admin añade una canción a un album"
      operationId: "add_album_song"
      consumes:
      - "application/json"
      produces:
      - "application/json"
      parameters:
      - name: "albumID"
        in: "path"
        description: "ID del album"
        required: true
        type: "string"
      - in: "body"
        name: "songID"
        description: "Song to add"
        required: false
        schema:
          type: "string"
          format: "int32"
      responses:
        201:
          description: "song added"
          schema:
            $ref: "#/definitions/AlbumItem"
        400:
          description: "invalid input, object invalid"
        401:
          description: "need authentification"
        403:
          description: "operation forbidden for user"
        404:
          description: "not found"
      x-swagger-router-controller: "swagger_server.controllers.admins_controller"
...
  /authors/{authorID}/image:
    get:
      tags:
      - "public"
      summary: "obtiene la imagen de un autor"
      description: "Obtiene la imagen de un autor identificado por authorID\n"
      operationId: "get_author_image"
      produces:
      - "image/png"
      parameters:
      - name: "authorID"
        in: "path"
        description: "ID del autor"
        required: true
        type: "string"
      responses:
        200:
          description: "success"
          schema:
            type: "file"
        400:
          description: "bad input parameter"
        404:
          description: "not found"
      x-swagger-router-controller: "swagger_server.controllers.public_controller"
    post:
      tags:
      - "admins"
      summary: "carga la imagen e un autor"
      description: "Un admin carga la imagen de un autor con identificador authorID."
      operationId: "upload_author_image"
      consumes:
      - "multipart/form-data"
      produces:
      - "application/json"
      parameters:
      - name: "authorID"
        in: "path"
        description: "ID del autor"
        required: true
        type: "string"
      - name: "authorImage"
        in: "formData"
        description: "Image item to update"
        required: true
        type: "file"
      responses:
        201:
          description: "file uploaded"
        400:
          description: "invalid input, object invalid"
        401:
          description: "need authentification"
        403:
          description: "operation forbidden for user"
        404:
          description: "not found"
      x-swagger-router-controller: "swagger_server.controllers.admins_controller"
    delete:
      tags:
      - "admins"
      summary: "elimina la imagen de un autor"
      description: "Elimina la imagen de un autor identificado por authorID\n"
      operationId: "delete_author_image"
      produces:
      - "application/json"
      parameters:
      - name: "authorID"
        in: "path"
        description: "ID del autor"
        required: true
        type: "string"
      responses:
        200:
          description: "image deleted"
        400:
          description: "bad input parameter"
        401:
          description: "need authentification"
        403:
          description: "operation forbidden for user"
        404:
          description: "not found"
      x-swagger-router-controller: "swagger_server.controllers.admins_controller"
...
definitions:
  SongItemNew:
    type: "object"
    required:
    - "albumID"
    - "authorID"
    - "genre"
    - "lenght"
    - "name"
    properties:
      name:
        type: "string"
        example: "Nombre de la nueva canción"
      lenght:
        type: "string"
        format: "int32"
        example: "03:05"
      authorID:
        type: "string"
        format: "int32"
        example: "6789"
      albumID:
        type: "string"
        format: "int32"
        example: "45"
      genre:
        type: "array"
        items:
          type: "string"
          example: "Géneros de la nueva canción"
    example:
      lenght: "03:05"
      name: "Nombre de la nueva canción"
      genre:
      - "Géneros de la nueva canción"
      - "Géneros de la nueva canción"
      albumID: "45"
      authorID: "6789"
  SongItem:
    type: "object"
    required:
    - "albumID"
    - "albumName"
    - "authorID"
    - "authorName"
    - "genre"
    - "id"
    - "lenght"
    - "name"
    properties:
      id:
        type: "string"
        format: "int32"
        example: "123456789"
      name:
        type: "string"
        example: "I Will Survive"
      lenght:
        type: "string"
        format: "int32"
        example: "03:05"
      authorID:
        type: "string"
        format: "int32"
        example: "4567"
      authorName:
        type: "string"
        example: "Gloria Gaynor"
      albumID:
        type: "string"
        format: "int32"
        example: "9876"
      albumName:
        type: "string"
        example: "Allenrok"
      genre:
        type: "array"
        items:
          type: "string"
          example: "rock"
    example:
      lenght: "03:05"
      albumName: "Allenrok"
      authorName: "Gloria Gaynor"
      name: "I Will Survive"
      genre:
      - "rock"
      - "rock"
      albumID: "9876"
      id: "123456789"
      authorID: "4567"
...
  AlbumItemNew:
    type: "object"
    required:
    - "authorID"
    - "name"
    - "publishDate"
    properties:
      name:
        type: "string"
        example: "Nombre del album"
      authorID:
        type: "string"
        format: "int32"
        example: "9876"
      publishDate:
        type: "string"
        format: "date"
        example: "2017-07-21T00:00:00.000Z"
      description:
        type: "string"
        example: "Descripción del album"
    example:
      name: "Nombre del album"
      publishDate: "2017-07-21T00:00:00.000Z"
      description: "Descripción del album"
      authorID: "9876"
  AlbumItem:
    type: "object"
    required:
    - "authorID"
    - "authorName"
    - "description"
    - "id"
    - "name"
    - "songs"
    properties:
      id:
        type: "string"
        format: "int32"
        example: "98765"
      name:
        type: "string"
        example: "Allenrok"
      authorID:
        type: "string"
        format: "int32"
        example: "4678"
      authorName:
        type: "string"
        example: "Estopa"
      publishDate:
        type: "string"
        format: "date"
        example: "2017-07-21T00:00:00.000Z"
      description:
        type: "string"
        example: "A tope, nene"
      songs:
        type: "array"
        items:
          $ref: "#/definitions/SongItem"
    example:
      authorName: "Estopa"
      songs:
      - lenght: "03:05"
        albumName: "Allenrok"
        authorName: "Gloria Gaynor"
        name: "I Will Survive"
        genre:
        - "rock"
        - "rock"
        albumID: "9876"
        id: "123456789"
        authorID: "4567"
      - lenght: "03:05"
        albumName: "Allenrok"
        authorName: "Gloria Gaynor"
        name: "I Will Survive"
        genre:
        - "rock"
        - "rock"
        albumID: "9876"
        id: "123456789"
        authorID: "4567"
      name: "Allenrok"
      publishDate: "2017-07-21T00:00:00.000Z"
      description: "A tope, nene"
      id: "98765"
      authorID: "4678"
...